% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_setup.R
\name{verify_setup}
\alias{verify_setup}
\title{Verify system setup}
\usage{
verify_setup()
}
\value{
Returns \code{TRUE} if assigned. Otherwise produces status messages.
}
\description{
Call this function to ensure that all necessary packages are installed.
}
